+++ webpack.config.babel.js
@@ -40,9 +40,26 @@ const config = {
     },
   ],
   module: {
-    loaders: [
-      {test: /\.js$/, exclude: /node_modules/, loader: 'babel'},
-    ],
+    rules: [{
+      test: /\.js$/,
+      exclude: /node_modules/,
+      loader: 'babel-loader',
+      options: {
+        plugins: [
+          'dev-expression',
+          'lodash',
+          'transform-proto-to-assign',
+        ],
+        presets: [
+          ['es2015', {
+            loose: true,
+            modules: false,
+          }],
+          'react',
+          'stage-1',
+        ],
+      },
+    }],
   },
 }
 
@@ -71,6 +88,9 @@ export default [
       path: outputPath,
     },
     plugins: [
+      new webpack.LoaderOptionsPlugin({
+        minimize: true,
+      }),
       new webpack.DefinePlugin({
         'process.env.NODE_ENV': JSON.stringify('production'),
       }),